+++ perllib/Open311/GetServiceRequestUpdates.pm
@@ -304,17 +304,20 @@ sub comment_text_for_request {
     my $ext_code_changed = $ext_code ne $old_ext_code;
     my $template;
     if ($state_changed || $ext_code_changed) {
+        my $order;
         my $state_params = {
             'me.state' => $state
         };
         if ($ext_code) {
             $state_params->{'me.external_status_code'} = $ext_code;
+            # make sure that empty string/nulls come last.
+            $order = { order_by => \"me.external_status_code DESC NULLS LAST" };
         };
 
         if (my $t = $problem->response_templates->search({
             auto_response => 1,
             -or => $state_params,
-        })->first) {
+        }, $order )->first) {
             $template = $t->text;
         }
     }