+++ CHANGELOG.md
@@ -12,6 +12,8 @@
         - Move summary failures to a separate script.
         - Add script to export/import body data.
         - Add fetch script that does combined job of fetch-comments and fetch-reports.
+    - Open311 improvements:
+        - match response templates on external status code over state
     - UK:
         - Added junction lookup, so you can search for things like "M60, Junction 2"